+++ package.json
@@ -12,6 +12,7 @@
     "bootstrap": "^5.3.2",
     "nodemon": "^3.0.1",
     "react-bootstrap": "^2.9.1",
+    "react-router-dom": "^6.18.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
