+++ src/App.js
@@ -1,4 +1,6 @@
+import {BrowserRouter, Routes, Route} from 'react-router-dom'
 import Home from './components/Home';
+import RenderNews from './components/RenderNews';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
 
@@ -6,7 +8,12 @@ import './App.css';
 function App() {
   return (
     <div className="App">
-      <Home />
+      <BrowserRouter>
+      <Routes>        
+        <Route path='/*' element={<Home />} >
+        (<Route path=':id' element={<RenderNews  />} />) 
+        </Route></Routes>
+      </BrowserRouter>
     </div>
   );
 }
